Move ListProperty into a new package
