+++ src/main/java/fr/ourten/teabeans/value/list/ArrayListProperty.java
@@ -1,4 +1,4 @@
-package fr.ourten.teabeans.value;
+package fr.ourten.teabeans.value.list;
 
 import java.util.List;
